General changes to JavaScript, CSS, HTML
Removed the jQuery link at the bottom of the HTML file because it was broken. Removed the jQuery code that was in the JavaScript file and replaced it with window.location.reload();
Made general changes to the styling of various elements.
Formatted each file to make the code easier to read.