General changes to all files
Adding divs  to the html file to separate the elements out more.  Removed the inline styling for the drop down menu. Added a description and keywords meta tags in the head of the HTML file.  Added a icon image.

For the CSS file I did a ground up rework of the styling and added a grid lay out. Changed styling of the background, buttons, input field, separating line, list section, and remove buttons. Removed the bullet point from appearing at the begging of each list item.

For the JavaScript file I re-ordered how the list items appear so that the remove button is at the beginning of each list item. I feel like it looks better.  Added a "enter" key event listener to the input field so the user can tap enter on the keyboard in stead of having to click the add item button.